src/main/java/Coordinate.java
import java.util.ArrayList;

public class Coordinate {
    private int x;
    private int y;

    public Coordinate(int x, int y) {
        this.x = x;
        this.y = y;
    }

    public int getX() {return x;}
    public int getY() {return y;}

    @Override
    public boolean equals(Object o) {
        if (o == this) {
            return true;
        }

        if (!(o instanceof Coordinate)) {
            return false;
        }

        Coordinate oCoord = (Coordinate) o;

        return x == oCoord.getX() && y == oCoord.getY();
    }
}
